DB/Schema: Boost stats in `player_levelstats` from tinyint to smallint
Needed to accommodate base stats higher than 255

Ref #13823

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -3222,7 +3222,7 @@ void ObjectMgr::LoadPlayerInfo()
 
                 PlayerLevelInfo& levelInfo = info->levelInfo[current_level - 1];
                 for (int i = 0; i < MAX_STATS; i++)
-                    levelInfo.stats[i] = fields[i + 3].GetUInt8();
+                    levelInfo.stats[i] = fields[i + 3].GetUInt16();
             }
 
             ++count;
